clean up - remove extra modules directory
